move responsibility build archive.Config Run to caller
