+++ cmd/slack-archive-lambda/main.go
@@ -1,0 +1,9 @@
+package main
+
+import archive "github.com/ToshihitoKon/slack-archive"
+
+func main() {
+	// TODO: impl
+	conf := &archive.Config{}
+	archive.Run(conf)
+}
