+++ interface.go
@@ -2,7 +2,6 @@ package archive
 
 import (
 	"context"
-	"time"
 )
 
 /* Example
@@ -30,34 +29,4 @@ type TextExporterInterface interface {
 }
 type FileExporterInterface interface {
 	WriteFiles(context.Context, []*LocalFile, func(*LocalFile) string) error
-}
-
-type LocalFile struct {
-	id        string
-	path      string
-	name      string
-	timestamp time.Time
-}
-
-type Output struct {
-	ID        string    `json:"id,omitempty"`
-	Timestamp time.Time `json:"timestamp,omitempty"`
-	Username  string    `json:"username,omitempty"`
-	Text      string    `json:"text,omitempty"`
-
-	Replies    Outputs `json:"replies,omitempty"`
-	LocalFiles []*LocalFile
-}
-
-type Outputs []*Output
-
-func (outputs Outputs) LocalFiles() []*LocalFile {
-	res := []*LocalFile{}
-	for _, output := range outputs {
-		res = append(res, output.LocalFiles...)
-		for _, reply := range output.Replies {
-			res = append(res, reply.LocalFiles...)
-		}
-	}
-	return res
 }
